Create new IconPropsArray and update Objective with it
New class IconPropsArray.cs was created under ScriptableOject/IconProps. This class acts as a container storing an array of IconProps. The Objective.cs script was updated to link to this new IconPropsArray instead of direct IconProps. This change is intended to streamline the handling of IconProps in the future, making it easier to manage multiple IconProps.
Also, the Unity layout has changed from Project to Scene and the current layout had some minor updates.
The EmployeeWorker.cs was also updated to work with these new changes.
This commit will make the project structure cleaner and more manageable for future developments.