Update project settings and sprites attributes
In this commit, several attribute changes were made to URP-Balanced.asset, URP-HighFidelity.asset, URP-Performant.asset, and other files to improve the performance and graphical quality of the game. Changes include enhanced screen space occlusion, updated shadow settings, optimization of the renderer, etc.

The texture sizes of several sprites were also reduced to reduce the memory utilization, and their compression was modified to improve efficiency.

Changes in the ObjectiveDisplay.cs file were made for prettier design.

Several enhancements were also done to the BurstAot and CommonBurstAot settings to increase the performance of the game.

A LevelUp check was added to EmployeeWorker.cs to fix a missing requirement, which will improve the gameplay experience.