"Add Menu related classes and prefabs, update scenes to reflect these changes
Created classes MenuClose and MenuSetup, included in the Scripts/Menu directory. These classes control the behavior of menu closing and setting up respectively.
Added prefabs to be used in menus - 'PostItMenuClose Variant'. This variant can be used to create a UI button to close menus.
Updated the GamePreferences to manage volume of background music.
Modified MenuScene and DesktopScene Unity files to accommodate these changes.
These changes were necessary to implement a more structured approach to managing menus in the game."