Add new Calculator and Code mini-games
Added two new mini-game objects, Calculator and Code, to the MenuScene. The Calculator game is intended to help employee characters resolve problems by playing mini-games. The purpose of the Code game is to indicate that an employee has encountered a problem or a bug.

These changes are part of an effort to increase user engagement by introducing new gameplay elements.