Add swipe sound effect and improve menu usability
Added 'swip.wav' sound for game transitions and made it playable at the game initiation. Improved the visibility of the menu, removing the 'menuHint' objects when a mini game instance is initiated.

In the RubiksColorHandler, HashWordHandler and MiniGameManager scripts, set the 'menuHint' game object to inactive on enabling the mini game. This change was implemented to make the game menu less cluttered and more user-friendly.

Optimized the quality of game scenes by adjusting shadow bias and color grading mode